* cukey/cushi; 8 bit keycodes

+++ shared/jaykay/cushi.h
@@ -1,3 +1,3 @@
 #pragma once
 
-uint16_t process_cushi_keys(uint16_t keycode, keyrecord_t *record, bool simulate);
+uint8_t process_cushi_keys(uint8_t keycode, keyrecord_t *record, bool simulate);
